update zklink scripts
