+++ hardhat.base.config.js
@@ -1,7 +1,6 @@
 require("@nomiclabs/hardhat-solpp");
 
 require("./script/deploy_zklink");
-require("./script/upgrade_zklink");
 require("./script/deploy_lz_bridge");
 require("./script/deploy_faucet");
 require("./script/deploy_account_mock");
